set method dispatches on multiple types
Since we use getText to extract a json string from the jsoneditor
control, it should only be expected that giving it a string would be
interpreted as a string with json in it. Because a string it itself
json, however, not switching on the type causes it to insert the
string literal into the editor. Even though this is a bit awkward, I
think this behavior is more predictable.